+++ flutter_local_notifications/android/src/main/java/com/dexterous/flutterlocalnotifications/FlutterLocalNotificationsPlugin.java
@@ -1,5 +1,6 @@
 package com.dexterous.flutterlocalnotifications;
 
+import static android.provider.Settings.ACTION_MANAGE_APP_USE_FULL_SCREEN_INTENT;
 import static android.provider.Settings.ACTION_REQUEST_SCHEDULE_EXACT_ALARM;
 
 import android.Manifest;
@@ -163,6 +164,9 @@ public class FlutterLocalNotificationsPlugin
       "requestNotificationsPermission";
   private static final String REQUEST_EXACT_ALARMS_PERMISSION_METHOD =
       "requestExactAlarmsPermission";
+
+  private static final String REQUEST_FULL_SCREEN_INTENT_PERMISSION_METHOD =
+      "requestFullScreenIntentPermission";
   private static final String METHOD_CHANNEL = "dexterous.com/flutter/local_notifications";
   private static final String INVALID_ICON_ERROR_CODE = "invalid_icon";
   private static final String INVALID_LARGE_ICON_ERROR_CODE = "invalid_large_icon";
@@ -204,6 +208,8 @@ public class FlutterLocalNotificationsPlugin
   static final int NOTIFICATION_PERMISSION_REQUEST_CODE = 1;
 
   static final int EXACT_ALARM_PERMISSION_REQUEST_CODE = 2;
+
+  static final int FULL_SCREEN_INTENT_PERMISSION_REQUEST_CODE = 3;
 
   private PermissionRequestListener callback;
 
@@ -1452,6 +1458,20 @@ public class FlutterLocalNotificationsPlugin
               }
             });
         break;
+      case REQUEST_FULL_SCREEN_INTENT_PERMISSION_METHOD:
+        requestFullScreenIntentPermission(
+            new PermissionRequestListener() {
+              @Override
+              public void complete(boolean granted) {
+                result.success(granted);
+              }
+
+              @Override
+              public void fail(String message) {
+                result.error(PERMISSION_REQUEST_IN_PROGRESS_ERROR_CODE, message, null);
+              }
+            });
+        break;
       case PERIODICALLY_SHOW_METHOD:
         repeat(call, result);
         break;
@@ -1835,6 +1855,36 @@ public class FlutterLocalNotificationsPlugin
     }
   }
 
+  public void requestFullScreenIntentPermission(@NonNull PermissionRequestListener callback) {
+    if (permissionRequestProgress != PermissionRequestProgress.None) {
+      callback.fail(PERMISSION_REQUEST_IN_PROGRESS_ERROR_MESSAGE);
+      return;
+    }
+
+    this.callback = callback;
+
+    if (Build.VERSION.SDK_INT >= VERSION_CODES.UPSIDE_DOWN_CAKE) {
+      NotificationManager notificationManager =
+          (NotificationManager) applicationContext.getSystemService(Context.NOTIFICATION_SERVICE);
+      AlarmManager alarmManager = getAlarmManager(applicationContext);
+      boolean permissionGranted = notificationManager.canUseFullScreenIntent();
+
+      if (!permissionGranted) {
+        permissionRequestProgress = PermissionRequestProgress.RequestingFullScreenIntentPermission;
+        mainActivity.startActivityForResult(
+            new Intent(
+                ACTION_MANAGE_APP_USE_FULL_SCREEN_INTENT,
+                Uri.parse("package:" + applicationContext.getPackageName())),
+            FULL_SCREEN_INTENT_PERMISSION_REQUEST_CODE);
+      } else {
+        this.callback.complete(true);
+        permissionRequestProgress = PermissionRequestProgress.None;
+      }
+    } else {
+      this.callback.complete(true);
+    }
+  }
+
   @Override
   public boolean onRequestPermissionsResult(
       int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
@@ -2167,7 +2217,8 @@ public class FlutterLocalNotificationsPlugin
   @Override
   public boolean onActivityResult(int requestCode, int resultCode, @Nullable Intent data) {
     if (requestCode != NOTIFICATION_PERMISSION_REQUEST_CODE
-        && requestCode != EXACT_ALARM_PERMISSION_REQUEST_CODE) {
+        && requestCode != EXACT_ALARM_PERMISSION_REQUEST_CODE
+        && requestCode != FULL_SCREEN_INTENT_PERMISSION_REQUEST_CODE) {
       return false;
     }
 
@@ -2176,6 +2227,15 @@ public class FlutterLocalNotificationsPlugin
         && VERSION.SDK_INT >= VERSION_CODES.S) {
       AlarmManager alarmManager = getAlarmManager(applicationContext);
       this.callback.complete(alarmManager.canScheduleExactAlarms());
+      permissionRequestProgress = PermissionRequestProgress.None;
+    }
+
+    if (permissionRequestProgress == PermissionRequestProgress.RequestingFullScreenIntentPermission
+        && requestCode == FULL_SCREEN_INTENT_PERMISSION_REQUEST_CODE
+        && VERSION.SDK_INT >= VERSION_CODES.UPSIDE_DOWN_CAKE) {
+      NotificationManager notificationManager =
+          (NotificationManager) applicationContext.getSystemService(Context.NOTIFICATION_SERVICE);
+      this.callback.complete(notificationManager.canUseFullScreenIntent());
       permissionRequestProgress = PermissionRequestProgress.None;
     }
 
@@ -2200,6 +2260,7 @@ public class FlutterLocalNotificationsPlugin
   enum PermissionRequestProgress {
     None,
     RequestingNotificationPermission,
-    RequestingExactAlarmsPermission
+    RequestingExactAlarmsPermission,
+    RequestingFullScreenIntentPermission
   }
 }